Fixing end of wizard back into app and removing consent from main app
