+++ app/views/consent.scala.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="../shared/build/bridge-shared.css" />
     <link rel="stylesheet" href="consent/build/consent.css" />
 </head>
-<body ng-app="consent" ng-controller="MainController" ng-cloak>
+<body ng-app="consent" ng-controller="ConsentController" ng-cloak>
 
 <div id="welcome" ng-show="step == 'welcome'">
     <header>
