--- public/app/views/consent.html
@@ -1,8 +1,0 @@
-<h3>Consent to Participate in Our Study</h3>
-
-<p>This would be the consent form, whatever that eventually looks like for the desktop consent process. </p>
-
-<div class="topSpacer">
-	<button ng-click="agree()" class="btn btn-primary">I agree to participate</button>
-	<button ng-click="decline()" class="btn btn-warning">I do not agree</button>
-</div>
