+++ public/consent/styles/consent.scss
@@ -194,6 +194,19 @@ article {
         font-size: 1.8rem;
     }
 }
+#thankyou {
+    h3 {
+        font-size: 2.2rem;
+    }
+    .info {
+        font-size: 1.8rem;
+    }
+    .sub_info {
+        margin-top: 10rem;
+        text-align: center;
+        font-size: 1.8rem;
+    }
+}
 .flipped {
     -webkit-transform: rotate(-180deg);
     -moz-transform: rotate(-180deg);
